backoffice/requirements.txt
django-allauth==0.43.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-timezone-field==4.1.1
Pillow==8.0.1
oauthlib==3.1.0
django-simple-history==3.0.0